pyproject.toml
[tool.poetry]
name = "familiarity-playlist-backend"
version = "0.1.0"
description = ""
authors = ["Pony0w0 <pony8674312@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "=3.10.13"
django = "4.2.13"
django-redis = "5.0.0"
djangorestframework = "3.15.2"
django-filter = "24.3"
django-extensions = "4.1"
psycopg2-binary = "2.9.6"
python-dotenv = "1.0.0"
django-cors-headers = "4.6.0"
ipython = "8.10"
gunicorn = "23.0.0"
djangorestframework-simplejwt = "5.4.0"
requests = "2.31.0"
celery = {version = "5.5.3", extras = ["redis"]}
django-celery-beat = "2.8.1"
paho-mqtt = "1.6.1"
tenacity = "8.2.3"
django-postgres-extra = "2.0.9"
django-import-export = "4.3.9"
cryptography = "45.0.3"
django-phonenumber-field = {version = "7.3.0", extras = ["phonenumbers"]}
channels = "4.2.0"
channels-redis = "4.2.0"
daphne = "4.1.2"
gdal = "3.6.2"
pyproj = "3.4.1"
pillow = "10.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
